requirements.txt
bg-news-trf @ https://huggingface.co/sakelariev/bg_news_trf/resolve/main/bg_news_trf-3.5.4-py3-none-any.whl
cuda-pathfinder==1.3.4
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl
langcodes==3.5.1
pathy==0.11.0
pipdeptree==2.30.0
presidio-analyzer==2.2.35
PyMuPDF==1.27.1
torchaudio==2.8.0.dev20250911+cu128
torchvision==0.24.0.dev20250911+cu128
wrapt==2.1.1
